feat: system-consolidation-validation-test-file-exemption-burn-down-003 - Preserve invocation contract validation evidence
